src/requirements.txt
streamlit
rembg
pillow
opencv-python